workspaces/sasoli0502/staff-task-system/pyproject.toml
[project]
name = "staff-task-system"
version = "0.1.0"
description = "スタッフタスク管理システム - 日次業務タスクの割り振りと進捗管理"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.uv]
package = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []
